Refactoring: introduce handler for messages
