--- ProjectV/WebServices/ProjectV.TelegramBotWebService/v1/Domain/Commands/IBotCommandHandler.cs
@@ -1,6 +1,0 @@
-﻿namespace ProjectV.TelegramBotWebService.v1.Domain.Commands
-{
-    public interface IBotCommandHandler
-    {
-    }
-}
