+++ ProjectV/WebServices/ProjectV.TelegramBotWebService/v1/Domain/Handlers/IBotHandler.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace ProjectV.TelegramBotWebService.v1.Domain.Handlers
+{
+    public interface IBotHandler<TType> : IDisposable
+    {
+        Task ProcessAsync(TType type);
+    }
+}
